feat: Bug fixes & Test Recipes (#373)

+++ test/e2e/context.ts
@@ -1,30 +1,28 @@
-import { spawn } from 'child_process';
+import { ChildProcess, spawn, spawnSync } from 'child_process';
+import { rmSync } from 'fs';
 import { join } from 'path';
-import tmpdir = require('temporary-directory');
 
-import { ProcessStatus } from './process';
 import { TestServer } from './server';
+import { createLogger } from './utils';
 
-/** A temporary directory handle. */
-interface TempDirectory {
-  /** Absolute path to the directory. */
-  path: string;
-  /** A function that will remove the directory when invoked. */
-  cleanup(): void;
+function getDeleteDirectories(appName: string): string[] {
+  switch (process.platform) {
+    case 'win32':
+      return [
+        join(process.env.APPDATA || '', appName),
+        join(process.env.LOCALAPPDATA || '', 'Temp', `${appName} Crashes`),
+      ];
+    case 'darwin':
+      return [join('~/Library/Application Support', appName)];
+    case 'linux':
+      return [join(process.env.XDG_CONFIG_HOME || '~/.config', appName)];
+  }
+
+  throw new Error('Unknown platform');
 }
 
-/** Creates a temporary directory with a cleanup function. */
-async function getTempDir(): Promise<TempDirectory> {
-  return new Promise<TempDirectory>((resolve, reject) => {
-    tmpdir((err, dir, cleanup) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve({ cleanup, path: dir });
-      }
-    });
-  });
-}
+const log = createLogger('Test Context');
+const appLog = createLogger('App');
 
 if (!process.env.DEBUG) {
   // tslint:disable-next-line
@@ -35,12 +33,6 @@ if (!process.env.DEBUG) {
 export class TestContext {
   /** Can check if the main process is running and kill it */
   public mainProcess?: ProcessStatus;
-
-  /** App stdout used for writing to console on failed tests */
-  public processStdOut: string = '';
-
-  /** Temporary directory that hosts the app's User Data. */
-  private _tempDir?: TempDirectory;
 
   private _started: boolean = false;
 
@@ -53,50 +45,45 @@ export class TestContext {
    */
   public constructor(
     private readonly _electronPath: string,
-    private readonly _appPath: string = join(__dirname, 'test-apps', 'node-integration'),
+    private readonly _appPath: string,
+    private readonly _appName: string,
   ) {}
 
   /** Starts the app. */
-  public async start(sentryConfig?: string, fixture?: string): Promise<void> {
-    // Only setup the tempDir if this the first start of the context
-    // Subsequent starts will use the same path
-    if (!this._tempDir) {
-      // Get a temp directory for this app to use as userData
-      this._tempDir = await getTempDir();
-    }
+  public async start(options: { secondRun?: boolean } = {}): Promise<void> {
+    log('Starting test context');
 
-    const env: { [key: string]: string | undefined } = {
+    const env: Record<string, any | undefined> = {
       ...process.env,
-      DSN: 'http://37f8a2ee37c0409d8970bc7559c7c7e4@localhost:8123/277345',
-      E2E_TEST_SENTRY: sentryConfig,
-      E2E_USERDATA_DIRECTORY: this._tempDir.path,
       ELECTRON_ENABLE_LOGGING: process.env.DEBUG,
+      ELECTRON_DISABLE_SECURITY_WARNINGS: true,
     };
 
-    if (fixture) {
-      env.E2E_TEST_FIXTURE = fixture;
+    if (!options.secondRun) {
+      env.APP_FIRST_RUN = true;
     }
 
     const childProcess = spawn(this._electronPath, [this._appPath], { env });
 
-    // eslint-disable-next-line no-extra-boolean-cast
-    if (!!process.env.DEBUG) {
-      childProcess.stdout.pipe(process.stdout);
-      childProcess.stderr.on('data', (data) => {
-        const str = data.toString();
-        if (str.match(/^\[\d+:\d+/)) {
-          return;
-        }
-        process.stderr.write(data);
-      });
-    } else {
-      childProcess.stdout.on('data', (data) => {
-        this.processStdOut += data.toString();
-      });
-      childProcess.stderr.on('data', (data) => {
-        this.processStdOut += data.toString();
-      });
+    function logLinesWithoutEmpty(input: string): void {
+      input
+        // Replace all the lines from the renderer
+        .replace(/^\[\d+:\d+\S+] "([\s\S]+?)"(?:[\s\S]+?)$/gm, (_, msg) => {
+          return `[Renderer] ${msg
+            .split(/[\r\n]+/)
+            .filter((e: string) => e.match(/\S/))
+            .join('\r\n[Renderer] ')}`;
+        })
+        .split(/[\r\n]+/)
+        // ignore empty lines
+        .filter((e: string) => e.match(/\S/))
+        // Add [Main] to all non renderer lines
+        .map((e: string) => (e.startsWith('[Renderer]') ? e : `[    Main] ${e}`))
+        .forEach((e: string) => appLog(e));
     }
+
+    childProcess.stdout.on('data', (data) => logLinesWithoutEmpty(data.toString()));
+    childProcess.stderr.on('data', (data) => logLinesWithoutEmpty(data.toString()));
 
     this.mainProcess = new ProcessStatus(childProcess);
 
@@ -105,19 +92,25 @@ export class TestContext {
       'Timeout: Waiting for app to start',
     );
 
+    log('App process has started');
+
     this._started = true;
   }
 
   /** Stops the app and cleans up. */
-  public async stop(clearData: boolean = true): Promise<void> {
-    if (!this.mainProcess) {
-      throw new Error('Invariant violation: Call .start() first');
+  public async stop(options: { retainData?: boolean } = {}): Promise<void> {
+    log('Stopping test context');
+
+    if (!this.mainProcess || !this.isStarted) {
+      throw new Error('Invariant violation: Call start() first');
     }
 
     await this.mainProcess.kill();
 
-    if (this._tempDir && clearData) {
-      this._tempDir.cleanup();
+    if (!options.retainData) {
+      for (const dir of getDeleteDirectories(this._appName)) {
+        rmSync(dir, { recursive: true, force: true });
+      }
     }
   }
 
@@ -130,7 +123,7 @@ export class TestContext {
   public async waitForTrue(
     method: () => boolean | Promise<boolean>,
     message: string = 'Timeout',
-    timeout: number = 8000,
+    timeout: number = 8_000,
   ): Promise<void> {
     if (!this.mainProcess) {
       throw new Error('Invariant violation: Call .start() first');
@@ -143,6 +136,7 @@ export class TestContext {
       await new Promise<void>((resolve) => setTimeout(resolve, 100));
       remaining -= 100;
       if (remaining < 0) {
+        log(message);
         throw new Error(message);
       }
     }
@@ -154,11 +148,56 @@ export class TestContext {
    *
    * @param count Number of events to wait for
    */
-  public async waitForEvents(testServer: TestServer, count: number, timeout: number = 15000): Promise<void> {
-    await this.waitForTrue(() => testServer.events.length >= count, 'Timeout: Waiting for events', timeout);
+  public async waitForEvents(testServer: TestServer, count: number, timeout: number = 15_000): Promise<void> {
+    log(`Waiting for ${count} events`);
+    await this.waitForTrue(
+      () => testServer.events.length >= count,
+      `Timeout: Waiting ${timeout}ms for ${count} events. Only ${testServer.events.length} events received`,
+      timeout,
+    );
+    log(`${count} events received`);
   }
 
-  public isStarted(): boolean {
+  /** Waits for app to close */
+  public async waitForAppClose(): Promise<void> {
+    await this.waitForTrue(
+      async () => (this.mainProcess ? !(await this.mainProcess.isRunning()) : false),
+      'Timeout: Waiting for app to die',
+    );
+  }
+
+  public get isStarted(): boolean {
     return this._started;
+  }
+}
+
+/** Handle for a running process. */
+export class ProcessStatus {
+  public constructor(private readonly _chProcess: ChildProcess) {}
+
+  /** Kills the process if it is still running. */
+  public async kill(): Promise<void> {
+    if (await this.isRunning()) {
+      this._chProcess.kill();
+    }
+
+    if (process.platform == 'win32') {
+      // The tests sometimes hang in CI on Windows because the Electron processes don't exit
+      spawnSync('taskkill /F /IM electron.exe', { shell: true });
+    }
+  }
+
+  /** Determines whether this process is still running. */
+  public async isRunning(): Promise<boolean> {
+    try {
+      if (this._chProcess.pid) {
+        process.kill(this._chProcess.pid, 0);
+        return true;
+      }
+    } catch (e) {
+      //
+    }
+
+    return false;
   }
 }
